Move protocol models to common module
