--- mvicore-plugin-middleware/build.gradle
@@ -1,67 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'maven'
-apply plugin: 'kotlin'
-apply plugin: 'org.jetbrains.dokka'
-
-group = 'com.github.badoo.mvicore'
-
-dependencies {
-    def deps = rootProject.ext.deps
-
-    implementation deps('io.reactivex.rxjava2:rxjava')
-    implementation deps('io.reactivex.rxjava2:rxkotlin')
-    implementation deps('com.google.code.gson:gson')
-    implementation project(':mvicore')
-
-    testImplementation deps('junit:junit')
-    testImplementation deps('org.jetbrains.kotlin:kotlin-test-junit')
-    testImplementation deps('org.amshove.kluent:kluent')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-}
-
-sourceCompatibility = "1.8"
-targetCompatibility = "1.8"
-
-repositories {
-    jcenter()
-}
-
-sourceSets {
-    main {
-        java {}
-    }
-}
-
-task packageSources(type: Jar, dependsOn: 'classes') {
-    classifier = 'sources'
-    from sourceSets.main.allSource
-}
-
-task packageJavadoc(type: Jar, dependsOn: javadoc) {
-    from javadoc.outputDirectory
-    classifier = 'javadoc'
-}
-
-artifacts {
-    archives packageSources
-    archives packageJavadoc
-}
-buildscript {
-    ext.kotlin_version = '1.3.21'
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-}
-compileKotlin {
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-compileTestKotlin {
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
