--- mvicore-plugin/build.gradle
@@ -1,60 +1,0 @@
-plugins {
-    id "org.jetbrains.intellij" version "0.4.7"
-}
-
-apply plugin: 'java'
-apply plugin: 'kotlin'
-apply plugin: 'org.jetbrains.intellij'
-apply plugin: 'idea'
-
-group = 'com.github.badoo'
-version = '0.0.1'
-
-intellij {
-    pluginName 'mvicore-plugin'
-    version "2019.1"
-    plugins = ['android']
-}
-
-patchPluginXml {
-    sinceBuild '131'
-}
-
-dependencies {
-    def deps = rootProject.ext.deps
-
-    implementation deps('io.reactivex.rxjava2:rxjava')
-    implementation deps('io.reactivex.rxjava2:rxkotlin')
-    implementation deps('com.google.code.gson:gson')
-
-    testImplementation deps('junit:junit')
-    testImplementation deps('org.jetbrains.kotlin:kotlin-test-junit')
-    testImplementation deps('org.amshove.kluent:kluent')
-}
-
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-repositories {
-    jcenter()
-
-}
-
-task packageSources(type: Jar, dependsOn: 'classes') {
-    classifier = 'sources'
-    from sourceSets.main.allSource
-}
-
-compileKotlin {
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-artifacts {
-    archives packageSources
-}
-
-jar {
-    from configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
-}
