+++ mvicore-plugin/common/.gitignore
