+++ mvicore-plugin/common/src/main/java/com/badoo/mvicore/plugin/model/Event.kt
@@ -1,0 +1,24 @@
+package com.badoo.mvicore.plugin.model
+
+import com.google.gson.JsonElement
+
+sealed class Event(val type: String) {
+    data class Bind(
+        val connection: ConnectionData
+    ) : Event("bind")
+
+    data class Item(
+        val connection: ConnectionData,
+        val element: JsonElement
+    ) : Event("data")
+
+    data class Complete(
+        val connection: ConnectionData
+    ) : Event("complete")
+
+    data class Connect(
+        val name: String
+    ) : Event("connect")
+
+    object Ping: Event("ping")
+}
