+++ mvicore-plugin/idea/.gitignore
