+++ mvicore-plugin/idea/src/main/java/com/badoo/mvicore/plugin/ui/JsonTreeNode.kt
