+++ mvicore-plugin/idea/src/main/java/com/badoo/mvicore/plugin/utils/Adb.kt
