+++ mvicore-plugin/idea/src/main/resources/META-INF/plugin.xml
