+++ mvicore-plugin/middleware/.gitignore
@@ -1,0 +1,1 @@
+/build
