+++ mvicore-plugin/middleware/src/main/java/com/badoo/mvicore/middleware/gson/MviPluginTypeAdapterFactory.kt
