--- mvicore-plugin/src/main/java/com/badoo/mvicore/plugin/model/Item.kt
@@ -1,8 +1,0 @@
-package com.badoo.mvicore.plugin.model
-
-import com.google.gson.JsonElement
-
-data class Item(
-    val connection: Connection,
-    val element: JsonElement
-)
