+++ settings.gradle
@@ -2,8 +2,9 @@ include ':mvicore'
 include ':mvicore-diff'
 include ':mvicore-android'
 include ':mvicore-debugdrawer'
-include ':mvicore-plugin-middleware'
-include ':mvicore-plugin'
+include ':mvicore-plugin:middleware'
+include ':mvicore-plugin:idea'
+include ':mvicore-plugin:common'
 include ':mvicore-demo:mvicore-demo-catapi'
 include ':mvicore-demo:mvicore-demo-feature1'
 include ':mvicore-demo:mvicore-demo-feature2'
